requirement.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
Pillow==9.2.0
pytz==2022.5
sqlparse==0.4.3
tzdata==2022.5
django-rest-authtoken
uritemplate
pyyaml